requirements.txt
matplotlib==3.8.1
minio==7.1.17
numpy==1.26.1
pandas==2.0.3
pymongo==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.1
seaborn==0.13.0
tensorboard==2.15.1
torch==2.1.0
typing_extensions==4.8.0
uvicorn==0.23.2
